Closes #2626: nits for Rx thread BeforeClass helpers
